Move Domain and Relation from hirm.hh to separate files.
